requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
astor==0.8.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
flake8==3.9.2
idna==3.2
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
six==1.16.0
testfixtures==6.18.1
toml==0.10.2
urllib3==1.26.6
